full refactoring of TEF and fuchsia backends

+++ src/subscriber/time_.dummy.ml
@@ -1,1 +1,1 @@
-let[@inline] get_time_ns () : float = 0.
+let[@inline] get_time_ns () : int64 = 0L
